move streamx calloption to internal package
