+++ internal/streamx/streamxclient/callopt.go
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2024 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package streamxclient
+
+import "context"
+
+// This file contains the CallOptions related structs and functions
+// because they are not ready to be exposed to users.
+
+// StreamCloseCallback define close callback of the stream
+type StreamCloseCallback func(err error)
+
+// CallOptions define stream call level options
+type CallOptions struct {
+	StreamCloseCallback []StreamCloseCallback
+}
+
+// Apply call options
+func (copts *CallOptions) Apply(opts []CallOption) {
+	for _, opt := range opts {
+		opt.f(copts)
+	}
+}
+
+// CallOption define stream call level option
+type CallOption struct {
+	f func(o *CallOptions)
+}
+
+type ctxKeyCallOptions struct{}
+
+// NewCtxWithCallOptions register CallOptions into context
+func NewCtxWithCallOptions(ctx context.Context) (context.Context, *CallOptions) {
+	copts := new(CallOptions)
+	return context.WithValue(ctx, ctxKeyCallOptions{}, copts), copts
+}
+
+// GetCallOptionsFromCtx get CallOptions from context
+func GetCallOptionsFromCtx(ctx context.Context) *CallOptions {
+	v := ctx.Value(ctxKeyCallOptions{})
+	if v == nil {
+		return nil
+	}
+	copts, ok := v.(*CallOptions)
+	if !ok {
+		return nil
+	}
+	return copts
+}
+
+// WithStreamCloseCallback register StreamCloseCallback
+func WithStreamCloseCallback(callback StreamCloseCallback) CallOption {
+	return CallOption{f: func(o *CallOptions) {
+		o.StreamCloseCallback = append(o.StreamCloseCallback, callback)
+	}}
+}
